update lat and long types to Double

+++ src/main/java/com/trackit/Device/Device.java
@@ -160,10 +160,10 @@ public class Device {
     private Integer fuelRailAbsolutePressure;
 
 
-    private Float lat;
+    private Double lat;
 
     @JsonProperty("long")
 
-    private Float lon;
+    private Double lon;
 
 }
